Add files via upload
This is the assignment for creating a form . the Points were given by you on the basis of which the form is created. The first page is a small brief about coding ( basically a webpage on which we were supposed to link the form). The second page is the form which pops in the different tab after we click the "Contact and Registration Form" . The third page is the acknowledgement that the content is saved. It pops out when we click the submit button on the form.  The code is provided to you in 'Formasiign2.html' that is the code of the brief webpage(pg1 ).Sorry i could not provide the whole code. I thought when clicked on it would display the page but it displaces the code. Sorry for the inconvenience.  Hope you like my project. 
Thanks and Regards, 
Aarya Mulik